+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## 0.0.2
-- add [Term.destruct2]
+- add `Term.destruct2`
+- add `Pattern.var_any`
 - comment fixes
 
 ## 0.0.1